requirements-dev.txt
-r requirements.txt
pytest>=8.0,<9
pytest-django>=4.9,<5
pytest-cov>=6.0,<7
